fix(formatter): preserve JetBrains diagnostic links
