+++ crates/oxc_diagnostics/src/lib.rs
@@ -47,6 +47,7 @@
 //! ```
 
 mod service;
+mod source_name;
 
 use std::{
     borrow::Cow,
@@ -56,7 +57,8 @@ use std::{
 
 pub mod reporter;
 
-pub use crate::service::{DiagnosticSender, DiagnosticService, to_file_url};
+pub use crate::service::{DiagnosticSender, DiagnosticService};
+pub use crate::source_name::with_file_url;
 
 pub type Error = miette::Error;
 pub type Severity = miette::Severity;
